fix: extract insertPromises
