+++ scripts/scrappers/src/updateProdDb.ts
@@ -2,7 +2,7 @@ import dotenv from "dotenv";
 import yargs from "yargs";
 import { gPromisesFromFiles } from "./helpers";
 import { GPromisesRepository } from "@mauriciorobayo/gods-promises/lib/repositories";
-import { IGPromise } from "../../../core/lib/models";
+import { insertGPromises } from "./insertPromises";
 
 dotenv.config();
 
@@ -33,28 +33,3 @@ dotenv.config();
 
   await insertGPromises(gPromises, gPromisesRepository);
 })();
-
-export async function insertGPromises(
-  gPromises: IGPromise[],
-  gPromisesRepository: GPromisesRepository
-) {
-  try {
-    const result = await gPromisesRepository.insertMany(gPromises, {
-      ordered: false,
-    });
-    console.log({ insertedCount: result.insertedCount });
-    process.exit();
-  } catch (err) {
-    if (err.code === 11000) {
-      const { result } = err.result;
-      console.log({
-        ok: result.ok,
-        writeErrors: result.writeErrors.length,
-        insertedIds: result.insertedIds.length,
-      });
-      process.exit();
-    }
-    console.error(err);
-    process.exit(1);
-  }
-}
